Fixed compile warnings / errors.
git-svn-id: https://srv183.svn-repos.de/dev34/pdclib/trunk@182 bcf39385-58cc-4174-9fcf-14f50f90dd47

+++ functions/stdio/gets.c
@@ -7,13 +7,13 @@
 */
 
 #include <stdio.h>
-#include <stdint.h>
+#include <limits.h>
 
 #ifndef REGTEST
 
 char * gets( char * s )
 {
-    return fgets( s, SIZE_MAX, stdin );
+    return fgets( s, INT_MAX, stdin ); /* TODO: Replace with an unchecking call. */
 }
 
 #endif
